requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.48
asyncpg==0.29.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6
redis==5.3.1
celery==5.3.6
psycopg2-binary==2.9.9
alembic==1.13.3